Port to 0.8.9.1

+++ src/main/kotlin/com/atsuishio/superbwarfare/entity/projectile/SmallCannonShellEntity.kt
@@ -1,10 +1,12 @@
 package com.atsuishio.superbwarfare.entity.projectile
 
+import com.atsuishio.superbwarfare.Mod.Companion.loc
 import com.atsuishio.superbwarfare.config.server.ExplosionConfig
 import com.atsuishio.superbwarfare.init.ModDamageTypes.causeProjectileHitDamage
 import com.atsuishio.superbwarfare.init.ModItems
 import com.atsuishio.superbwarfare.init.ModSounds
 import com.atsuishio.superbwarfare.network.message.receive.ClientIndicatorMessage
+import com.atsuishio.superbwarfare.resource.model.ProjectileModelReloadListener
 import com.atsuishio.superbwarfare.tools.CustomExplosion
 import com.atsuishio.superbwarfare.tools.forceHurt
 import com.atsuishio.superbwarfare.tools.sendPacketTo
@@ -15,9 +17,7 @@ import net.minecraft.world.entity.EntityType
 import net.minecraft.world.entity.LivingEntity
 import net.minecraft.world.entity.projectile.Projectile
 import net.minecraft.world.item.Item
-import net.minecraft.world.level.Explosion
 import net.minecraft.world.level.Level
-import net.minecraft.world.level.block.BellBlock
 import net.minecraft.world.level.entity.EntityTypeTest
 import net.minecraft.world.phys.BlockHitResult
 import net.minecraft.world.phys.EntityHitResult
@@ -26,9 +26,9 @@ import net.minecraft.world.phys.Vec3
 open class SmallCannonShellEntity(type: EntityType<out SmallCannonShellEntity>, level: Level) :
     FastThrowableProjectile(type, level) {
     private var aa = false
+    open val modelInstance = ProjectileModelReloadListener.getModel(MODEL)?.createInstance()
 
     init {
-        this.noCulling = true
         this.damageValue = 40f
         this.explosionDamageValue = 80f
         this.explosionRadiusValue = 5f
@@ -38,48 +38,28 @@ open class SmallCannonShellEntity(type: EntityType<out SmallCannonShellEntity>, 
         return ModItems.SMALL_SHELL_AP
     }
 
-    override fun onHitEntity(result: EntityHitResult) {
-        super.onHitEntity(result)
-        val entity = result.entity
-        val owner = this.owner
-        if (owner != null && owner.vehicle != null && entity == owner.vehicle) return
+    override fun afterHitEntity(result: EntityHitResult) {
         if (this.level() is ServerLevel) {
-            entity.forceHurt(causeProjectileHitDamage(this.level().registryAccess(), this, owner), damageValue)
-
-            if (entity is LivingEntity) {
-                entity.invulnerableTime = 0
-            }
-
             if (this.tickCount > 0) {
-                causeExplode(result.getLocation(), true)
+                this.causeExplode(result.getLocation(), true)
             }
-            this.discard()
         }
+        this.discard()
     }
 
-    public override fun onHitBlock(result: BlockHitResult) {
-        super.onHitBlock(result)
+    override fun afterHitBlock(result: BlockHitResult) {
         val resultPos = result.blockPos
-        val state = this.level().getBlockState(resultPos)
-
         if (this.level() is ServerLevel) {
             val hardness = this.level().getBlockState(resultPos).block.defaultDestroyTime()
             if (hardness != -1f) {
-                if (ExplosionConfig.EXPLOSION_DESTROY.get() && ExplosionConfig.EXTRA_EXPLOSION_EFFECT.get()) {
+                if (ExplosionConfig.EXPLOSION_DESTROY.get() && ExplosionConfig.EXTRA_EXPLOSION_EFFECT.get() && this.explosionDestroyValue) {
                     val destroy = Math.random() < (1.0 - (hardness / 50.0)).coerceIn(0.1, 1.0)
                     if (destroy) {
                         this.level().destroyBlock(resultPos, true)
                     }
                 }
             }
-        }
-
-        val block = state.block
-        if (block is BellBlock) {
-            block.attemptToRing(this.level(), resultPos, result.direction)
-        }
-        if (this.level() is ServerLevel) {
-            causeExplode(result.getLocation(), false)
+            this.causeExplode(result.getLocation(), false)
         }
         this.discard()
     }
@@ -90,9 +70,9 @@ open class SmallCannonShellEntity(type: EntityType<out SmallCannonShellEntity>, 
             .damage(explosionDamageValue)
             .radius(explosionRadiusValue)
             .position(vec3)
-            .withParticleType(explosionParticleType(explosionRadiusValue))
-            .destroyBlock { if (hitEntity) Explosion.BlockInteraction.KEEP else (if (ExplosionConfig.EXPLOSION_DESTROY.get()) Explosion.BlockInteraction.DESTROY else Explosion.BlockInteraction.KEEP) }
-            .damageMultiplier(1.25f)
+            .beast(this.isBeast())
+            .destroyBlock(if (hitEntity) false else explosionDestroyValue)
+            .beast(this.isBeast())
             .explode()
     }
 
@@ -109,7 +89,7 @@ open class SmallCannonShellEntity(type: EntityType<out SmallCannonShellEntity>, 
         }
     }
 
-    fun crushProjectile(velocity: Vec3) {
+    open fun crushProjectile(velocity: Vec3) {
         if (this.level() is ServerLevel) {
             val frontBox = boundingBox.inflate(0.5).expandTowards(velocity)
 
@@ -156,5 +136,9 @@ open class SmallCannonShellEntity(type: EntityType<out SmallCannonShellEntity>, 
 
     override fun isFastMoving(): Boolean {
         return false
+    }
+
+    companion object {
+        val MODEL = loc("models/bedrock/projectile/small_cannon_shell.geo.json")
     }
 }
